+++ inhibitors/permissions.js
@@ -5,6 +5,7 @@ exports.conf = {
 };
 
 exports.run = (client, msg, cmd) => {
-  if (msg.author.permLevel >= cmd.conf.permLevel) return false;
+  if (msg.channel.type !== "text" && msg.author.permLevel >= cmd.conf.permLevel) return false;
+  else if (msg.member.permLevel >= cmd.conf.permLevel) return false;
   return "You do not have permission to use this command.";
 };